requirements.txt
sounddevice
soundfile
numpy
sanic
python-socketio
pyserial
requests
